% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.axis.r
\name{freq.axis}
\alias{freq.axis}
\title{Function to determine y axis values}
\usage{
freq.axis(maxval)
}
\arguments{
\item{maxval}{Maximum value of frequency in original data frame}
}
\description{
Function to determine y axis values
}
