% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmap.r
\name{genmap}
\alias{genmap}
\title{Plot genetic/linkage map with all requested settings}
\usage{
genmap(
  linmap = NULL,
  lingroup = NULL,
  title = "",
  linkage.width = 0.4,
  label.x = "Distance (cM)",
  label.chr = NULL,
  cex.label.chr = 0.7,
  marker.color = "deepskyblue3",
  marker.count.color = "deepskyblue3",
  marker.count.width = 5,
  draw.border = TRUE,
  highlight.marker = NULL,
  highlight.marker.color = "red",
  show.axis = FALSE,
  show.marker.count = FALSE,
  cex.marker.count = 0.7,
  border = FALSE,
  border.linetype = "solid",
  border.linewidth = 1,
  border.color = "black",
  ...
)
}
\arguments{
\item{linmap}{Data frame of all marker plus linkage groups and frequency values}

\item{lingroup}{Character vector specifying the linkaga groups to draw}

\item{title}{Title of the plot}

\item{linkage.width}{Relative width of a single linkage group on plot (Default: 0.4)}

\item{label.x}{Text label of the x axis (Default: "Distance (cM)")}

\item{label.chr}{Text label shown beneath the linkage groups}

\item{cex.label.chr}{Size of the linkage group text labels (Default: 0.7)}

\item{marker.color}{Color of marker (Default: "deepskyblue")}

\item{marker.count.color}{Color of the marker count annotation (Default: "deepskyblue")}

\item{marker.count.width}{Width of bar of marker count annotation}

\item{draw.border}{Boolean value if border around linkage group should be displayed (Default: TRUE)}

\item{highlight.marker}{Character vector of all markers to highlight}

\item{highlight.marker.color}{Color of highlighted marker (Default: "red")}

\item{show.axis}{Boolean value if y axis should be displayed}

\item{show.marker.count}{Boolean value if marker count annotation should be displayed}

\item{cex.marker.count}{Size of text label of marker count annotation}

\item{border}{Boolean value if border around plot area should be displayed (Default: FALSE)}

\item{border.linetype}{Linetype of border line}

\item{border.linewidth}{Value of line width of border around plot}

\item{border.color}{Color of border around plot}

\item{...}{further parameter of plot function (base R)}
}
\description{
Plot genetic/linkage map with all requested settings
}
