% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.group.position.r
\name{linkage.group.position}
\alias{linkage.group.position}
\title{Calculate position of linkage group on plot area}
\usage{
linkage.group.position(plotSize = NULL, count = NULL)
}
\arguments{
\item{plotSize}{Dimension (size) of x axis}

\item{count}{Number of linkage groups}
}
\value{
positional coordinates
}
\description{
Calculate position of linkage group on plot area
}
