% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmap.r
\docType{data}
\name{linmap}
\alias{linmap}
\title{Linkage map dataset}
\usage{
data(linmap)
}
\description{
Table of linkage groups and markers and their corresponding recombination frequency (cM).
}
\details{
Arbitrarily created data of a genetic/linkage map
}
