% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placements.r
\name{placements}
\alias{placements}
\title{Conclude position of linkage group on plot area}
\usage{
placements(dfori, chromosomes, xsize, ysize, telsize)
}
\arguments{
\item{dfori}{Original data frame}

\item{chromosomes}{Character vector of requested linkage groups}

\item{xsize}{X size of plot area}

\item{ysize}{Y size of plot area}

\item{telsize}{Value of telomere size (Default = 40)}
}
\value{
Modified data frame extended by positional values
}
\description{
Conclude position of linkage group on plot area
}
