% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.map.r
\name{read.map}
\alias{read.map}
\title{Function to read genetic/linkage map}
\usage{
read.map(file = NULL, ...)
}
\arguments{
\item{file}{Filepath of genetic/linkage map}
}
\value{
Correctly formatted genetic/linkage map
}
\description{
Function to read genetic/linkage map
}
\examples{
read.map("your/path/map.txt")
}
