% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.axis.r
\name{scale.axis}
\alias{scale.axis}
\title{Function to determine y axis values}
\usage{
\method{scale}{axis}(maxval)
}
\arguments{
\item{maxval}{Maximum value of frequency in original data frame}
}
\description{
Function to determine y axis values
}
\examples{
scale.axis(maxval)
}
