% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomere.r
\name{telomere}
\alias{telomere}
\title{Function to plot telomere on plot}
\usage{
telomere(location, x, y, r, start = 0, end = pi, nsteps = 1000)
}
\arguments{
\item{location}{Position to plot telomere at}

\item{x}{Circle parameter}

\item{y}{Circle parameter}

\item{r}{Radius of half circle}

\item{start}{Start value of calculation (Default: 0)}

\item{end}{End value of calculation (Default: pi)}

\item{nsteps}{Number od steps in between for calculation (Default: 1000)}
}
\description{
Function to plot telomere on plot
}
\examples{
telomere(location, x, y, r, start, end, nsteps)
}
